New features: #retrieve AJAX, #parse, #hideBanner
The .retrieve() method currently logs to the console in order to spy on
the response it grabs. This should get replaced with something else as
soon as possible.

This method will return a JSON object when (if) the AJAX request was
successful.

The .parse() method takes a callback function as an argument. When
combined with the .retrieve() method, you can parse the JSON returned by
.retrieve(). It returns an object with the urls for any *.mp4, *.ogv, or
*.webm videos that have been uploaded and associated with this
Squarespace page.

The .hideBanner() method has been initiated. We'll end up comparing this
with each of the default templates (just for coverage).