Remove smell, add support for Adirondack template
Use a function within the spec for #hideBanner in order to write the
tests more succinctly: pass a template name (plus any additonal
nesting) and a data-image attribute value to match against to #testHide in
order to test if an element has been hidden. Add an optional false value
if it should NOT be hidden.

Also, made the first jQuery selector more specific so that the store
pages in the Adirondack template wouldn't be picked up.