Begin working on tests to support a #replace method
The .replace() method should make use of the other BannerVideo methods
(and helper functions in util.js) to GET the JSON for the page, parse
it, and hide the current banner image. Then, it will need to insert the
banner video.

It might be a good idea to create an .insert() method as well.

This is a minimal implimentation for using the other methods that
currently exist, but it works synchronously. Since we are making a GET
request as the page loads, we should probably be doing this
asynchronously instead. I plan on taking a stab at ES6 Promises to do
that.